Add Multiplier

+++ src/main/kotlin/implementation/mode/Mode.kt
@@ -5,7 +5,7 @@ import implementation.ColorImplementation
 interface Mode {
 
     // Will be called from another thread
-    fun start(color: ColorImplementation, multiplier: Int)
+    fun start(color: ColorImplementation, multiplier: Float)
 
 
     // Will be call from main thread
